+++ libs/core/src/string.c
@@ -1,5 +1,5 @@
 #include "core_string.h"
 
-i32 string_cmp(String a, String b) { return mem_cmp(a, b); }
+i8 string_cmp(String a, String b) { return mem_cmp(a, b); }
 
 bool string_eq(String a, String b) { return mem_eq(a, b); }
